first

+++ src/main/java/com/example/demo/service/impl/FileServiceImpl.java
@@ -104,7 +104,7 @@ public class FileServiceImpl implements IFileService {
     }
 
     @Override
-    public String deleteFile(String id, Long id_file, Users user) throws Exception {
+    public String deleteFile(String id, Long id_file, Users user, boolean actor) throws Exception {
 
         Optional<File> optionalFile = fileRepository.findById(id_file);
         if (!optionalFile.isPresent()) {
@@ -112,22 +112,27 @@ public class FileServiceImpl implements IFileService {
         }
 
         File file = optionalFile.get();
-
-        if (file.getUser().getId() == user.getId()) {
-            if (file.getLink().equals(id)) {
-                googleFileManager.deleteFileOrFolder(id);
-                fileRepository.deleteById(id_file);
-                user.setMaxUpload(user.getMaxUpload() + file.getFileSize());
-                userServiceImpl.save(user);
-                return "File deleted successfully.";
+        if (file != null) {
+            if (file.getUser().getId() == user.getId()) {
+                if (file.getLink().equals(id)) {
+                    googleFileManager.deleteFileOrFolder(id);
+                    fileRepository.deleteById(id_file);
+                    user.setMaxUpload(user.getMaxUpload() + file.getFileSize());
+                    userServiceImpl.save(user);
+                    if (actor == true) {
+                        userServiceImpl.sendDelete(user, file);
+                    }
+                    return "File deleted successfully.";
+                } else {
+                    throw new NotFoundException("File not found");
+                }
             } else {
-                throw new NotFoundException("File not found");
+                throw new NotFoundException(
+                        "The post cannot be deleted. Only the author of the post has the permission to delete it.");
             }
         } else {
-            throw new NotFoundException(
-                    "The post cannot be deleted. Only the author of the post has the permission to delete it.");
+            throw new NotFoundException("File not found");
         }
-
     }
 
     @Override
